Change modifyExpression to return a new promlb.Matcher
In previous versions of the prometheus client, promlb.Matcher was just a struct that contained strings.

However, in recent versions of the client, promlb.Matcher contains a FastRegexMatcher instantiated within it, which is only instantiated when you call promlb.NewMatcher on a matcher whose type is a regex match.

As a result, simply modifying the promlb.Matcher struct fields results in a segfault on trying to parse a regex label selector using a new client, which is fixed with the change introduced here.