Use full error message for some direct proxy test cases
At the time of writing this commit, the apiContextHandler called on hijacking requests received by Prometheus Auth had the following logic that only returns the underlying error that was returned by a call, not the overall error:

https://github.com/rancher/prometheus-auth/blob/f78ab78c65a1afe5c94c90c36856fd3c0ee627ef/pkg/agent/http_api_context.go#L152-L158

However, since in direct proxying we do not pass anything through the API Context Handler, the errors returned in direct proxying are slightly different (wrapped) versions of the errors in someNamespaceToken test cases.

Therefore, this commit encodes this logical difference between the behavior of hijacked and proxied requests in a test.